requirements.txt
Django==4.2.16
djangorestframework==3.15.2
mysqlclient==2.2.4
requests==2.32.3
Pillow
django-cors-headers